pyproject.toml
[project]
name = "trading-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "Matthew Holman",email = "mholman000@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic (>=1.16.4,<2.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "pre-commit (>=4.2.0,<5.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "sqlalchemy-utils (>=0.41.2,<0.42.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "lxml (>=6.0.0,<7.0.0)"
]


[tool.poetry.group.dev.dependencies]
requests-mock = "^1.10.0"
httpx = ">=0.22.0"
pre-commit = ">=4.2.0,<5.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
